requirements.txt
streamlit
pandas
bibtexparser
sumy
transformers
torch
sentence-transformers
llama-cpp-python
faiss-cpu
PyMuPDF
huggingface_hub
